backend/requirements.txt
Django==5.0.1
djangorestframework==3.14.0
django-cors-headers==4.3.1
openai==1.61.0
python-dotenv==1.0.0
pypdf==4.0.1
psycopg2-binary==2.9.9
pgvector==0.2.4
langchain==0.1.9
langchain-community==0.0.38
langchain-core==0.1.52
python-multipart==0.0.6
tiktoken==0.5.2
python-docx==1.0.1  
openpyxl==3.1.2    